Model brand

+++ src/main/java/com/thoen/demoapi/controllers/BrandController.java
@@ -1,12 +1,18 @@
 package com.thoen.demoapi.controllers;
 
+import com.thoen.demoapi.dto.BrandDTO;
+import com.thoen.demoapi.exception.ApiException;
 import com.thoen.demoapi.models.Brand;
 import com.thoen.demoapi.services.BrandService;
+import com.thoen.demoapi.utils.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/brands")
@@ -16,22 +22,45 @@ public class BrandController {
     private BrandService brandService;
 
     @GetMapping
-    public List<Brand> getAllBrand(){
-        return brandService.getAllBrands();
+    public ResponseEntity<Map<String, Object>> getAllBrand(){
+        try {
+            List<Brand> brands = brandService.getAllBrands();
+            Map<String, Object> response = new HashMap<>();
+            if (brands.isEmpty()){
+                response.put("brands", brands);
+                response.put("message", "Successfully");
+                response.put("status", HttpStatus.OK.value());
+                return ResponseEntity.ok(response);
+            } else {
+                response.put("brands", brands);
+                response.put("message", "Brands retrieved successfully");
+                response.put("status", HttpStatus.OK.value());
+                return ResponseEntity.ok(response);
+            }
+        } catch (Exception e){
+            throw new ApiException(HttpStatus.NOT_FOUND, "Brands not found");
+        }
     }
-
-
     @PostMapping
-    public Brand createBrand(@RequestBody Brand brand){
-        return  brandService.createBrand(brand);
+    public ResponseEntity<?> createBrand(@RequestBody BrandDTO brandDTO){
+        Brand brand = Mapper.brandDTOToBrand(brandDTO);
+        Brand newBrand = brandService.createBrand(brand);
+        BrandDTO brandDTOResponse = Mapper.brandToBrandDTO(newBrand);
+        return ResponseEntity.ok(brandDTOResponse);
     }
-
 
     @GetMapping("/{id}")
-    public ResponseEntity<Brand> getBrandById(@PathVariable Integer id){
-        return brandService.getBrandById(id).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
+    public ResponseEntity<Map<String, Object>> getBrandById(@PathVariable Integer id){
+        return brandService.getBrandById(id)
+                .map(brand -> {
+                    Map<String, Object> response = new HashMap<>();
+                    response.put("message", "successfully");
+                    response.put("status", HttpStatus.OK.value());
+                    response.put("brand", brand);
+                    return ResponseEntity.ok(response);
+                })
+                .orElseThrow(() -> new ApiException(HttpStatus.NOT_FOUND, "Brand not found"));
     }
-
 
     @PutMapping("/update/{id}")
     public ResponseEntity<Brand> updateBrand(@PathVariable Integer id,@RequestBody Brand brandUpdate){
@@ -39,14 +68,21 @@ public class BrandController {
         if(updateBrand != null){
             return  ResponseEntity.ok(updateBrand);
         } else {
-            return ResponseEntity.notFound().build();
+            throw new ApiException(HttpStatus.NOT_FOUND, "Brand not found");
         }
     }
 
     @DeleteMapping("/delete/{id}")
-    public  ResponseEntity<Void> deleteBrand(@PathVariable Integer id){
-        brandService.deleteBrand(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<Map<String, Object>> deleteBrand(@PathVariable Integer id){
+        if (brandService.getBrandById(id).isPresent()){
+            brandService.deleteBrand(id);
+            Map<String, Object> response = new HashMap<>();
+            response.put("message", "Brand deleted successfully");
+            response.put("status", HttpStatus.OK.value());
+            return ResponseEntity.ok(response);
+        } else {
+            throw new ApiException(HttpStatus.NOT_FOUND, "Brand not found");
+        }
     }
